+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file. The format 
 
 ### Improved
 - Accordion chevron orientation and animation performance
+- Accordion can now constrain height with Surface
+
+### Fixed
+- Accordion constrained height updates when items expand
 
 ## [v0.8.0]
 ### Improved